finished filter extraction
